Refactor into modules to add stats module
