+++ __tests__/lib/polling/parser.ts
@@ -1,4 +1,4 @@
-import { parsePollMetadata } from '../../../lib/polling/parser';
+import { parsePollMetadata } from 'modules/polls/lib/parser';
 import pollJson327 from './poll-327.js';
 import pollJson431 from './poll-431.js';
 import fs from 'fs';
