+++ __tests__/lib/polling/validator.ts
@@ -1,4 +1,4 @@
-import { validateText } from '../../../lib/polling/validator';
+import { validateText } from 'modules/polls/lib/validator';
 import fs from 'fs';
 const pollMetadata = fs.readFileSync(__dirname + '/poll-431.md').toString();
