+++ __tests__/pages/executive.tsx
@@ -9,7 +9,7 @@ import {
   DEMO_ACCOUNT_TESTS
 } from '../helpers';
 import { ExecutiveOverview } from '../../pages/executive';
-import proposals from '../../mocks/proposals.json';
+import proposals from 'modules/executives/api/mocks/proposals.json';
 import { accountsApi } from 'stores/accounts';
 
 jest.mock('@theme-ui/match-media', () => {
