+++ components/header/VotingWeight.tsx
@@ -2,7 +2,7 @@ import { Flex, Text } from 'theme-ui';
 import useSWR from 'swr';
 import useAccountsStore from 'stores/accounts';
 import getMaker from 'lib/maker';
-import { getVotingWeightCopy } from 'lib/polling/getVotingWeightCopy';
+import { getVotingWeightCopy } from 'modules/polls/lib/getVotingWeightCopy';
 
 export default function VotingWeight(props): JSX.Element {
   const account = useAccountsStore(state => state.currentAccount);
