+++ modules/app/types/paginationOptions.d.ts
@@ -1,0 +1,5 @@
+export type PaginationOptions = {
+  page: number,
+  pageSize: number,
+  limit: number
+}
