+++ modules/polls/lib/getCategories.ts
