+++ modules/polls/lib/getVotingWeightCopy.ts
