+++ pages/api/stats/[address].ts
@@ -1,20 +1,18 @@
 import invariant from 'tiny-invariant';
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { parsePollsMetadata } from 'modules/polls/api/parsePollMetadata';
-import { getConnectedMakerObj } from 'lib/api/utils';
-import { isSupportedNetwork } from 'lib/maker';
+import { isSupportedNetwork } from 'lib/maker/index';
 import { DEFAULT_NETWORK } from 'lib/constants';
 import withApiHandler from 'lib/api/withApiHandler';
+import { fetchStats } from 'modules/stats/api/fetchStats';
+
 
 export default withApiHandler(async (req: NextApiRequest, res: NextApiResponse) => {
   const network = (req.query.network as string) || DEFAULT_NETWORK;
+  const address = req.query.address as string;
   invariant(isSupportedNetwork(network), `unsupported network ${network}`);
 
-  const maker = await getConnectedMakerObj(network);
-  const pollList = await maker.service('govPolling').getAllWhitelistedPolls();
-  const polls = await parsePollsMetadata(pollList);
-
+  const stats = await fetchStats(address, network);
   res.setHeader('Cache-Control', 's-maxage=15, stale-while-revalidate');
-  res.status(200).json(polls);
+  res.status(200).json(stats);
 });
