+++ pages/polling/review.tsx
@@ -10,7 +10,7 @@ import shallow from 'zustand/shallow';
 
 import { useBreakpointIndex } from '@theme-ui/match-media';
 import { isDefaultNetwork, getNetwork } from 'lib/maker';
-import { getPolls } from 'lib/api';
+import { getPolls } from 'modules/polls/api/fetchPolls';
 import { isActivePoll, findPollById } from 'lib/utils';
 import PrimaryLayout from 'components/layouts/Primary';
 import SidebarLayout from 'components/layouts/Sidebar';
